+++ src/Aks/Aks/Commands/SetAzureRmAks.cs
@@ -80,6 +80,9 @@ namespace Microsoft.Azure.Commands.Aks
         [Parameter(Mandatory = false, HelpMessage = "Whether to use use Uptime SLA.")]
         public SwitchParameter EnableUptimeSLA { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Whether to enalbe OIDC issuer feature.")]
+        public SwitchParameter EnableOidcIssuer { get; set; }
+
         private ManagedCluster BuildNewCluster()
         {
             BeforeBuildNewCluster();
@@ -409,16 +412,20 @@ namespace Microsoft.Azure.Commands.Aks
                     {
                         if (EnableUptimeSLA.ToBool())
                         {
-                            cluster.Sku = new ManagedClusterSKU(name: "Basic", tier: "Paid");
+                            cluster.Sku = new ManagedClusterSKU(name: "Base", tier: "Standard");
                         }
                         else
                         {
-                            cluster.Sku = new ManagedClusterSKU(name: "Basic", tier: "Free");
+                            cluster.Sku = new ManagedClusterSKU(name: "Base", tier: "Free");
                         }
                     }
                     if (this.IsParameterBound(c => c.AadProfile))
                     {
                         cluster.AadProfile = AadProfile;
+                    }
+                    if (EnableOidcIssuer.IsPresent)
+                    {
+                        cluster.OidcIssuerProfile = new ManagedClusterOIDCIssuerProfile(enabled: true);
                     }
                     SetIdentity(cluster);